refactor(lumx-icons): split utils modules
Fixes running `yarn postinstall` (icons build) before deps are installed
