+++ packages/lumx-icons/override/generate/utils/unicode-to-css-code.cjs
@@ -1,0 +1,4 @@
+/** Reverse of `cssCodeToUnicode` */
+const unicodeToCssCode = (unicode) => `\\${unicode.toString(16).toUpperCase()}`;
+
+module.exports = { unicodeToCssCode };
